+++ task-management/src/tasks/tasks.service.ts
@@ -14,8 +14,8 @@ export class TasksService {
     private taskRepository: TaskRepository
   ) {}
 
-  async getTaskById(id: string): Promise<Task> {
-    const task = this.taskRepository.findOne({ id });
+  async getTaskById(id: string, userId: string): Promise<Task> {
+    const task = this.taskRepository.findOne({ id, userId });
 
     if(!task) {
       throw new NotFoundException(`Task with id ${id} not found.`)
@@ -24,21 +24,22 @@ export class TasksService {
     return task;
   }
 
-  async getTasks(filterDto: GetTaskFilteredDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(userId: string, filterDto: GetTaskFilteredDto): Promise<Task[]> {
+    return this.taskRepository.getTasks(userId, filterDto);
   }
 
   async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
     return this.taskRepository.createTask(createTaskDto);
   }
 
-  async updateTask(id: string, taskData: UpdateTaskDto): Promise<Task> {
-    await this.taskRepository.update({ id }, { ...taskData });
-    return this.getTaskById(id);
+  async updateTask(id: string, userId: string, taskData: UpdateTaskDto): Promise<Task> {
+    await this.taskRepository.updateTask({ id, userId }, { ...taskData });
+
+    return this.getTaskById(id, userId);
   }
 
-  async deleteTask(id: string): Promise<DeleteResult> {
-    return this.taskRepository.delete({ id });
+  async deleteTask(id: string, userId: string): Promise<DeleteResult> {
+    return this.taskRepository.delete({ id, userId });
   }
 
 }
